my_blog/urba_web/requirements.txt
asgiref==3.5.0
Django==4.0.2
django-ckeditor==6.3.0
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
listview==0.1.0
Pillow==9.0.1
sqlparse==0.4.2
psycopg2-binary==2.8.5
